+++ Example/Common/ExampleUdpConfig.h
@@ -9,7 +9,7 @@
 EXTERN_C_BEGIN
 
     const char* ExampleUdpConfig_GetHost(void);
-    int         ExampleUdpConfig_GetPort(void);
+    uint16_t    ExampleUdpConfig_GetPort(void);
     void        ExampleUdpConfig_GetEndpoint(struct SolidSyslogEndpoint * endpoint);
     uint32_t    ExampleUdpConfig_GetEndpointVersion(void);
